requirements.txt
coverage==4.5.3
Django==2.2
djangorestframework==3.9.2
pytz==2019.1
sqlparse==0.3.0
